Fix math.log2(2**1023) by implementing the base-2 special case in rbigint.
This commit changes the rpython subdirectory, which it shouldn't because this
is the py3.3 branch.  But there already are differences in the rpython
subdirectory between default and py3.3, which are related to this change.
Unfortunately, all these changes can't just be applied to default, because it
would change sematics of the Python 2.7 implementation.  There should be a
discussion about how to refactor this.